LAB04/02-CloudAlbum-Chalice/cloudalbum/requirements.txt
Pillow==6.1.0
boto3==1.9.214
pynamodb==4.0.0
tzlocal==2.0.0
python-jose==3.0.1
requests==2.22.0
aws-parameter-store==1.0.0
requests-toolbelt==0.9.1
